git-svn-id: https://svn.met.no/diana/trunk@1181 0ccb8871-b436-41a3-ae7d-e3ac59973202

+++ include/diSat.h
@@ -123,7 +123,7 @@ public:
   int no;               ///< no of plotted channels
   int rgbindex[3];      ///< channelindex for rgb-operations
   unsigned char* rawimage[maxch]; ///< raw image
-  long* origimage[3]; ///< original image for temperature display images
+  float* origimage[3]; ///< original image for temperature display images
   int rawchannels[maxch];         ///< raw images channel numbers
   
   int calibidx;         ///< channel to use in values routine
